try cast

+++ c/driver/postgresql/postgres_copy_reader_test.cc
@@ -75,7 +75,7 @@ class PostgresCopyStreamWriteTester {
     // TODO: don't think we should do this here; the reader equivalent does
     // increment the data pointer and seemingly discard at the end, but
     // we may still want to keep that buffer available?
-    buffer->data -= writer_.array_size_approx_bytes();
+    buffer->data -= static_cast<uintptr_t>(writer_.array_size_approx_bytes());
     return result;
   }
